cours 05/05/21

+++ USAL38_K/Gestion_etudiant_V3/afficherEtudiant.php
@@ -1,9 +1,19 @@
 <?php
+// Vérification d'authentification validée
+require_once("verifAuth.php");
+require_once("connexionPDO.php");
 
-require_once ("connexion.php");
 
-$sql = "SELECT code, nom, photo FROM etudiant";
-$result = $conn->query($sql);
+// $sql = "SELECT code, nom, photo FROM etudiant";
+// $result = $conn->query($sql);
+
+$stmt = $conn->prepare("select code, nom, photo from etudiant");
+$stmt->bindParam(':code', $code);
+$stmt->bindParam(':nom', $nom);
+$stmt->bindParam(':photo', $photo);
+
+// Exécution de la requête sql
+$stmt->execute();
 
 /* if ($result->num_rows > 0) {
   // output data of each row
@@ -17,25 +27,32 @@ $result = $conn->query($sql);
 
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Suppression étudiant</title>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Afficher étudiant</title>
 </head>
+
 <body>
-    <table border="1" width="80%">
-        <tr>
-            <th>CODE</th><th>NOM</th><th>PHOTO</th>
-        </tr>
-        <?php while($row = $result->fetch_assoc()) { ?>
-            <tr>
-                <td><?php echo($row['code']) ?></td>
-                <td><?php echo($row['nom']) ?></td>
-                <td><img src="images/<?php echo($row['photo']) ?>"></td>
-                <td><a href="supprimerEtudiant.php?code=<?php echo($row['code']) ?>">Supprimer</a></td>
-                <td><a href="editerEtudiant.php?code=<?php echo($row['code']) ?>">Editer</a></td>
-            </tr>
+  <table border="1" width="80%">
+    <tr>
+      <th>CODE</th>
+      <th>NOM</th>
+      <th>PHOTO</th>
+    </tr>
+    <?php while ($row = $stmt->fetch( PDO::FETCH_ASSOC )) { ?>
+      <tr>
+        <td><?php echo ($row['code']) ?></td>
+        <td><?php echo(htmlspecialchars($row['nom'])) ?></td>
+        <td><img src="images/<?php echo ($row['photo']) ?>"></td>
+        <?php if ($_SESSION['NIVEAU'] == 0) { ?>
+          <td><a href="supprimerEtudiant.php?code=<?php echo ($row['code']) ?>">Supprimer</a></td>
+          <td><a href="editerEtudiant.php?code=<?php echo ($row['code']) ?>">Editer</a></td>
         <?php } ?>
-    </table>
+      </tr>
+    <?php } ?>
+  </table>
 </body>
+
 </html>
